Carrying Progress
Klonoa can now carry enemies around with his Bullet. Also made changes so that only Klonoa can get Dream stones, as well as fixing not getting hit when staying in an enemy's area. We're getting kloser to the true Klonoa experience! (That one was a mistake, but I'm keeping it LOL)